Update firmware.ino for smoother movement
I modified the pulse width to do more accurate "bit banging" as well as added more calls to each servo pulse since the servo doesn't fully get to the angle on one call. It seems after 30 calls we get as close as we are going to get to the desired angle.